LayoutTests/fast/canvas/pattern-too-large-to-create-expected.txt

CONSOLE MESSAGE: line 18: Canvas area exceeds the maximum limit (width * height > 2.68435e+8).
PASS: Saw exception.
